requirements.txt
torchvision~=0.9.1
matplotlib~=3.3.4
numpy~=1.20.2
tqdm~=4.59.0
pillow~=8.2.0